Remove duplicate `partitioned_by` check in dbt-athena's `incremental.sql`
In dbt-athena's `incremental.sql`, `partioned_by` was being checked twice.
First on line 33 where if `partioned_by` is `None` and the `strategy` is
`insert_overwrite`, then the strategy gets changed to `append`. Later on
line 66 we were again checking whether `partioned_by` was `None` in reference
to using the `insert_overwrite` strategy. However, due to the line 33 conditional,
the first part of the conditional on line 66 (checking `partioned_by`) would always
evaluate true if the strategy was still `insert_overwrite`.